Merge dev to main

+++ .assets/provision/install_exa.sh
@@ -27,29 +27,31 @@ SYS_ID=$(grep -oPm1 '^ID(_LIKE)?=.*\K(alpine|arch|fedora|debian|ubuntu|opensuse)
 
 case $SYS_ID in
 alpine)
-  apk add --no-cache exa >&2
+  apk add --no-cache exa >&2 2>/dev/null
   ;;
 arch)
-  pacman -Sy --needed --noconfirm exa >&2
+  pacman -Sy --needed --noconfirm exa >&2 2>/dev/null
   ;;
 fedora)
-  dnf install -y exa >&2
+  dnf install -y exa >&2 2>/dev/null
   ;;
 debian | ubuntu)
   export DEBIAN_FRONTEND=noninteractive
-  apt-get update >&2 && apt-get install -y exa >&2
+  apt-get update >&2 && apt-get install -y exa >&2 2>/dev/null
   ;;
 opensuse)
-  zypper in -y exa >&2
+  zypper in -y exa >&2 2>/dev/null
   ;;
-*)
+esac
+
+if ! type $APP &>/dev/null; then
+  echo 'Installing from binary.' >&2
   while [[ ! -f exa-linux-x86_64.zip ]]; do
     curl -Lsk -o exa-linux-x86_64.zip "https://github.com/ogham/exa/releases/download/v${REL}/exa-linux-x86_64-v${REL}.zip"
   done
-  unzip exa-linux-x86_64.zip >&2
-  mv -f bin/exa /usr/bin/exa >&2
-  mv -f man/* $(manpath | cut -d : -f 1)/man1 >&2
-  mv -f completions/exa.bash /etc/bash_completion.d >&2
+  unzip exa-linux-x86_64.zip &>/dev/null
+  mv -f bin/exa /usr/bin/exa &>/dev/null
+  mv -f man/* $(manpath | cut -d : -f 1)/man1 &>/dev/null
+  mv -f completions/exa.bash /etc/bash_completion.d &>/dev/null
   rm -fr bin completions man exa-linux-x86_64.zip
-  ;;
-esac
+fi
